+++ CHANGELOG.md
@@ -76,6 +76,8 @@
   This should improve support for some system without GUIs.
 * Refactored code in `augmenters.segmentation` (general code and docstring cleanup). #334
 * Refactored code in `augmenters.arithmetic` (general code and docstring cleanup). #328
+* Added check to `dtypes.gate_dtypes()` verifying that arguments `allowed`
+  and `disallowed` have no intersection.
 
 ## Fixes